global changes
Changes are associated with recognition capabilities.
Now the Sack is trained once and it is saved as an opportunity to rewrite the bag itself.
The mechanism of image recognition has not been completed yet. All vector data is read perfectly. It is necessary to add reading classes, to identify the class when recognizing.